+++ src/Auth.Application/Roles/Commands/Disabled/DisabledRoleCommand.cs
@@ -1,7 +1,7 @@
 ﻿using Auth.Application.Attributtes;
 using MediatR;
 
-namespace Auth.Application.Roles.Commands.Delete
+namespace Auth.Application.Roles.Commands.Disabled
 {
     [Authorize(AuthPermisions.RoleDisabled)]
     public class DisabledRoleCommand : IRequest
