+++ src/Auth.Application/Roles/Commands/Disabled/DisabledRoleCommandHandler.cs
@@ -6,7 +6,7 @@ using Microsoft.EntityFrameworkCore;
 using System.Threading;
 using System.Threading.Tasks;
 
-namespace Auth.Application.Roles.Commands.Delete
+namespace Auth.Application.Roles.Commands.Disabled
 {
     public class DisabledRoleCommandHandler : IRequestHandler<DisabledRoleCommand>
     {
